+++ pydash/__init__.py
@@ -435,6 +435,7 @@ from .utilities import (
     range_,
     range_right,
     result,
+    stub_string,
     times,
     to_path,
     unique_id,